Small style changes to prevent broken 'drilling down' on overview.
The intention was to provide teachers with a way to drill down into individual students from the overview page.

For reasons, this work had to be postponed.  In order not to confuse the teachers we removed any UI cues to suggest that summary table elements were clickable. Instead just the page headers are clickable.